# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: Archon API
  version: "1.0.0"
  description: Enterprise AI Orchestration & Governance Platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Primary API
security:
  - BearerAuth: []
paths:
...
  /agents:
    get:
      operationId: listAgents
      summary: List all agents
      tags:
        - agents
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createAgent
      summary: Create a new agent
      tags:
        - agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreate"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ────────────────────────────────────────────
  # Agents — single resource
  # ────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by Keycloak. Obtain a token via the Keycloak OpenID Connect
        token endpoint (grant_type=authorization_code or client_credentials).
        Include as `Authorization: Bearer <token>`.
    KeycloakOAuth2:
      type: oauth2
      description: Keycloak OAuth 2.0 authorization code and client credentials flows.
      flows:
        authorizationCode:
          authorizationUrl: /realms/archon/protocol/openid-connect/auth
          tokenUrl: /realms/archon/protocol/openid-connect/token
          refreshUrl: /realms/archon/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email address
            roles: User role assignments
        clientCredentials:
          tokenUrl: /realms/archon/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            roles: User role assignments
    OpenIDConnect:
      type: openIdConnect
      description: Keycloak OpenID Connect discovery endpoint.
      openIdConnectUrl: /realms/archon/.well-known/openid-configuration

  # ────────────────────────────────────────────
  # Parameters
  # ────────────────────────────────────────────
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad request — invalid or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "BAD_REQUEST"
                message: "Request body is malformed"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "UNAUTHORIZED"
                message: "Bearer token is missing or expired"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
...
    UnprocessableEntity:
      description: Validation error — input is well-formed but semantically invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "VALIDATION_ERROR"
                message: "Field value is invalid"
                field: "name"
                details: "Name must be between 1 and 255 characters"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "INTERNAL_ERROR"
                message: "An unexpected error occurred"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"

  # ────────────────────────────────────────────
  # Schemas
  # ────────────────────────────────────────────
  schemas:
...
    ResponseMeta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique identifier for this request (correlation ID)
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when the response was generated
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    PaginationMeta:
      type: object
      required:
        - total
        - limit
        - offset
        - has_more
      properties:
        total:
          type: integer
          description: Total number of items matching the query
          minimum: 0
        limit:
          type: integer
          description: Maximum items per page
          minimum: 1
          maximum: 100
        offset:
          type: integer
          description: Number of items skipped
          minimum: 0
        has_more:
          type: boolean
          description: Whether more items exist beyond this page
    ErrorEnvelope:
      type: object
      required:
        - errors
        - meta
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ErrorDetail"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
            - "VALIDATION_ERROR"
            - "NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Name of the field that caused the error (for validation errors)
        details:
          type: string
          description: Additional context about the error

    # ── Health ────────────────────────────────
...
    AgentCreate:
      type: object
      required:
        - name
        - definition
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable agent name
        description:
          type: string
          maxLength: 2000
          description: Optional description of the agent's purpose
        definition:
          type: object
          description: JSON graph definition describing the agent's node topology and edges
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 64
          maxItems: 20
          description: Optional tags for categorisation and search
        template_id:
          type: string
          format: uuid
          description: Optional ID of the template this agent was created from
...
    Agent:
      type: object
      required:
        - id
        - name
        - definition
        - status
        - owner_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique agent identifier
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable agent name
        description:
          type: string
          maxLength: 2000
          description: Description of the agent's purpose
        definition:
          type: object
          description: JSON graph definition (nodes, edges, parameters)
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
          description: Tags for categorisation and search
        template_id:
          type: string
          format: uuid
          description: ID of the template this agent was created from (if any)
        status:
          type: string
          enum: [draft, active, archived]
          description: Current lifecycle status of the agent
        owner_id:
          type: string
          format: uuid
          description: ID of the user who owns this agent
        created_at:
          type: string
          format: date-time
          description: Timestamp when the agent was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of the last update
    AgentResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    AgentListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    # ── Execution ─────────────────────────────
...